refact: Simplify CLI tool interactions in FlacConverter class
The wiring here seems needlessly complicated, so I took a step back and
made every effort to simplify it.

The fact of the matter is that the methods available in those classes,
e.g., SoxCommand, are very specific to SoX. They don't need to be
abstract enough to work with, say, a different CLI tool; they need only
be flexible enough to be impervious to changes *in the same CLI tool*,
such as when arguments or switches are changed, re-ordered, etc.