refact: Eliminate a great deal of redundancy and inefficiency
Most importantly, I addressed a TODO by pulling a process call outside
of a loop, and accomplishing the same task with a single process call.

Previously, a new process was started for each tag (i.e., each
key-value pair), and now, all of the pairs are added to a single
command prior to execution.