refact: Delete unused MP3 conversion code
Converting MP3 files to other formats isn't particularly useful
because MP3 is lossy and there are few practical use-cases for it,
excepting only instances in which a lossless source format is not
available.

If sufficient demand arises in the future, such capabilities could
readily be added, as it's conceivable that a user may want to
convert an archive of historical MP3s to OGG files, for example.