fix: Replace CLI tagging tool for MP3 (from lame to mid3v2)
Lame cannot tag-in-place; it re-encodes the file, which is not only
wasteful of resources, but a lossy process that further degrades the
quality of the resultant audio.

So, while Lame will still be used for transcoding (converting),
the ever-capable mid3v2 CLI tool will be used for tagging MP3 files.