fix: Forego using picture switch in favor of legacy artwork embedding
Only the newest versions of mid3v2 support the --picture switch,
whereas in older versions, it is necessary to target the APIC frame.
In the interest of supporting a wider variety of environments, at least
until specific binary versions are included as proper dependencies,
the legacy method has been employed.