feat: Add support for the latest getID3 version
getID3 v1.9.18 introduced a change, described as
'standardize "track" -> "track_number"', which broke
Mp3TaggingTest::testItCanTagFile(). In short, getID3 had previously returned
track numbers in the format 'tracknumber' => '1/10', whereas as of this
release, the format changed to 'tracknumber' => '1' *and* 'tracktotal' => '10'
was added.

This commit fixes the breaking change in getID3 in a backward-compatible way by
adding support for an optional mutator (passed as a closure) that manipulates
the tag value arbitrarily.