fix: Encoding failure when converting from FLAC to ALAC
Modern versions of ffmpeg (tested on stock Ubuntu 20.04 LTS with ffmpeg 4.2.4) throw an exception when attempting to convert a FLAC file to ALAC when the FLAC file contains a cover image in the video stream, unless specific CLI switches are passed.

Passing "-vn" will omit the cover image altogether, whereas passing "-vcodec copy" will preserve it; failure to do one or the other results in an exception.

The latter technique is mentioned in https://unix.stackexchange.com/questions/415477/lossless-audio-conversion-from-flac-to-alac-using-ffmpeg#comment1026896_415478

It is unknown why failing to pass the aforementioned switches did not cause problems in previous versions.